refactor mcp dto multiline tests
